Made moved files private
