+++ exasol/error/_error.py
@@ -8,13 +8,13 @@ from typing import (
     Union,
 )
 
-from exasol.error import exa_error
+from exasol.error import _exa_error
 from exasol.error._internal_errors import (
     INVALID_ERROR_CODE,
     LIBRARY_ERRORS,
     UNKNOWN_EXCEPTION_OCCURED,
 )
-from exasol.error.error_message_builder import (
+from exasol.error._error_message_builder import (
     ErrorMessageBuilder,
     InvalidErrorCode,
 )
