+++ exasol/error/_error_message_builder.py
@@ -6,8 +6,8 @@ from typing import (
     Optional,
 )
 
-from exasol.error.parameters_mapper import ParametersMapper
-from exasol.error.placeholder_handler import PlaceholderHandler
+from exasol.error._parameters_mapper import ParametersMapper
+from exasol.error._placeholder_handler import PlaceholderHandler
 
 ERROR_CODE_FORMAT = "^[FEW]-[A-Z][A-Z0-9]+(-[A-Z][A-Z0-9]+)*-[0-9]+$"
