+++ exasol/error/_placeholder.py
