+++ test/unit/legacy/test_error_code_format.py
@@ -2,7 +2,7 @@ import unittest
 
 import pytest
 
-from exasol.error.error_message_builder import (
+from exasol.error._error_message_builder import (
     ErrorMessageBuilder,
     InvalidErrorCode,
 )
